requirements.txt
telethon
ccxt
google-genai
aiosqlite
python-dotenv
